--- contract_comm/evm.go
@@ -1,26 +1,0 @@
-// Copyright 2017 The Celo Authors
-// This file is part of the celo library.
-//
-// The celo library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// The celo library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with the celo library. If not, see <http://www.gnu.org/licenses/>.
-
-package contract_comm
-
-import (
-	"github.com/celo-org/celo-blockchain/contract_comm/caller"
-	"github.com/celo-org/celo-blockchain/core/vm"
-)
-
-func SetInternalEVMHandler(chain vm.ChainContext) {
-	caller.SetInternalEVMHandler(chain)
-}
